+++ CHANGES.md
@@ -2,6 +2,35 @@
 
 ## 2019 09 15.01
 
+    - implement issue RT#130640: Allow different subroutine keywords.
+      Added a flag --sub-alias-list=s or -sal=s, where s is a string with
+      one or more aliases for 'sub', separated by spaces or commas.  
+      For example,
+
+        perltidy -sal='method fun' 
+
+      will cause the perltidy to treate the words 'method' and 'fun' to be
+      treated the same as if they were 'sub'.
+
+    - Added flag --space-prototype-paren=i, or -spp=i, to control spacing 
+      before the opening paren of a prototype, where i=0, 1, or 2:
+      i=0 no space
+      i=1 follow input [current and default]
+      i=2 always space
+
+      Previously, perltidy always followed the input.
+      For example, given the following input 
+
+         sub usage();
+
+      The result will be:
+        sub usage();    # i=0 [no space]
+        sub usage();    # i=1 [default; follows input]
+        sub usage ();   # i=2 [space]
+
+
+## 2019 09 15
+
     - fixed issue RT#130344: false warning "operator in print statement" 
       for "use lib".